Clarify Map logic 1
